+++ docs-gen/build-docs.py
@@ -21,7 +21,7 @@ book2.title = "Other Page"
 book2.chapters.append(book2.chapters[0])
 book2.chapters[1].sections.pop()
 
-pages = cl.compile_books(cl.codegen.HTMLCodegen, meta, book, book2)
+pages = cl.compile_books(cl.html.HTMLCodegen, meta, book, book2)
 for page in pages:
     #print(pages)
     pass
