+++ docs-gen/catlog/catlog.py
@@ -1,4 +1,4 @@
-from . import gml, doc, codegen
+from . import doc, gml, html, md, txt
 from pathlib import Path
 import os
 import re
