+++ docs-gen/catlog/md.py
@@ -1,0 +1,6 @@
+from .codegen import BasicCodegen
+
+class MarkdownCodegen(BasicCodegen):
+    EXT = ".md"
+
+    pass
