+++ docs-gen/catlog/txt.py
@@ -1,0 +1,6 @@
+from .codegen import BasicCodegen
+
+class TXTCodegen(BasicCodegen):
+    EXT = ".txt"
+
+    pass
